services/ocr_service.py
"""
OCR Service — Google Document AI with chunked PDF processing.

Google Document AI sync API limit: 15 pages per request.
For large PDFs (70-100 pages), we split into 14-page chunks,
process each, then concatenate all text with page markers.
"""
import os
import io
import json
import time
import logging
from dotenv import load_dotenv

load_dotenv()

logger = logging.getLogger(__name__)

GOOGLE_PROJECT_ID = os.getenv('GOOGLE_PROJECT_ID')
GOOGLE_PROCESSOR_ID = os.getenv('GOOGLE_PROCESSOR_ID')
GOOGLE_LOCATION = os.getenv('GOOGLE_LOCATION', 'us')
GOOGLE_APPLICATION_CREDENTIALS = os.getenv('GOOGLE_APPLICATION_CREDENTIALS')

OCR_RAW_FOLDER = os.getenv('OCR_RAW_FOLDER', 'storage/ocr/raw')
OCR_PROCESSED_FOLDER = os.getenv('OCR_PROCESSED_FOLDER', 'storage/ocr/processed')

# Safe page limit per chunk (Google Doc AI sync limit is 15)
CHUNK_SIZE = 14


def _resolve_credentials():
    """Resolve and set Google credentials path."""
    creds_path = GOOGLE_APPLICATION_CREDENTIALS
    if creds_path and not os.path.isabs(creds_path):
        project_root = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
        creds_path = os.path.join(project_root, creds_path)
    if creds_path and os.path.exists(creds_path):
        os.environ['GOOGLE_APPLICATION_CREDENTIALS'] = creds_path
        return creds_path
    # Auto-discover: look for any gothic*.json in project root
    project_root = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
    for f in os.listdir(project_root):
        if f.endswith('.json') and ('gothic' in f.lower() or 'service' in f.lower()):
            full = os.path.join(project_root, f)
            os.environ['GOOGLE_APPLICATION_CREDENTIALS'] = full
            return full
    return None


def _get_doc_ai_client():
    """Create and return a Document AI client."""
    _resolve_credentials()
    from google.cloud import documentai_v1 as documentai
    return documentai.DocumentProcessorServiceClient()


def _count_pdf_pages(pdf_path: str) -> int:
    """Return total page count using pypdf."""
    from pypdf import PdfReader
    reader = PdfReader(pdf_path)
    return len(reader.pages)


def _split_pdf_chunk(pdf_path: str, start_page: int, end_page: int) -> bytes:
    """
    Extract pages [start_page, end_page) from PDF and return as bytes.
    Zero-indexed pages.
    """
    from pypdf import PdfReader, PdfWriter
    reader = PdfReader(pdf_path)
    writer = PdfWriter()
    for i in range(start_page, min(end_page, len(reader.pages))):
        writer.add_page(reader.pages[i])
    buf = io.BytesIO()
    writer.write(buf)
    return buf.getvalue()


def _process_single_chunk(client, processor_name: str, pdf_bytes: bytes,
                           chunk_index: int, max_retries: int = 3) -> str:
    """
    Send one PDF chunk to Google Document AI.
    Retries up to max_retries times with exponential backoff.
    """
    from google.cloud import documentai_v1 as documentai

    raw_document = documentai.RawDocument(
        content=pdf_bytes,
        mime_type='application/pdf'
    )
    request_obj = documentai.ProcessRequest(
        name=processor_name,
        raw_document=raw_document
    )

    last_error = None
    for attempt in range(max_retries):
        try:
            result = client.process_document(request=request_obj)
            return result.document.text
        except Exception as e:
            last_error = e
            if attempt < max_retries - 1:
                wait = 2 ** attempt  # 1s, 2s
                logger.warning(
                    f"Chunk {chunk_index} attempt {attempt + 1} failed: {e}. "
                    f"Retrying in {wait}s…"
                )
                time.sleep(wait)

    raise RuntimeError(
        f"Chunk {chunk_index} failed after {max_retries} attempts: {last_error}"
    )


def extract_text_from_pdf(pdf_path: str, session_id: str,
                           progress_callback=None) -> str:
    """
    OCR a PDF of any size using chunked processing.

    - Counts total pages with pypdf
    - Splits into CHUNK_SIZE-page batches (14 pages each, safely under 15-page limit)
    - Processes each batch via Google Document AI with retry logic
    - Concatenates results with [PAGES x-y] markers so LLM knows page context
    - Saves raw metadata + processed text to disk
    - progress_callback(current, total) called after each chunk completes
    """
    os.makedirs(OCR_RAW_FOLDER, exist_ok=True)
    os.makedirs(OCR_PROCESSED_FOLDER, exist_ok=True)

    client = _get_doc_ai_client()
    from google.cloud import documentai_v1 as documentai

    processor_name = client.processor_path(
        GOOGLE_PROJECT_ID,
        GOOGLE_LOCATION,
        GOOGLE_PROCESSOR_ID
    )

    # Count pages
    total_pages = _count_pdf_pages(pdf_path)
    logger.info(f"PDF '{os.path.basename(pdf_path)}': {total_pages} pages")

    # Build chunk ranges
    chunk_ranges = [
        (start, min(start + CHUNK_SIZE, total_pages))
        for start in range(0, total_pages, CHUNK_SIZE)
    ]
    total_chunks = len(chunk_ranges)
    logger.info(f"Split into {total_chunks} chunk(s) of up to {CHUNK_SIZE} pages each")

    all_parts = []
    chunk_metadata = []

    for idx, (start, end) in enumerate(chunk_ranges):
        page_label = f"[PAGES {start + 1}–{end} of {total_pages}]"
        logger.info(f"  OCR chunk {idx + 1}/{total_chunks}: {page_label} — splitting PDF…")

        pdf_bytes = _split_pdf_chunk(pdf_path, start, end)
        logger.info(f"  OCR chunk {idx + 1}/{total_chunks}: {len(pdf_bytes):,} bytes → sending to Google Document AI…")
        chunk_text = _process_single_chunk(client, processor_name, pdf_bytes, idx)
        logger.info(f"  OCR chunk {idx + 1}/{total_chunks}: ✔ received {len(chunk_text):,} chars")

        # Add page marker for LLM context
        all_parts.append(f"\n{page_label}\n{chunk_text}")
        chunk_metadata.append({
            'chunk': idx + 1,
            'pages': f'{start + 1}-{end}',
            'chars_extracted': len(chunk_text),
        })

        if progress_callback:
            progress_callback(idx + 1, total_chunks)

    full_text = '\n'.join(all_parts)

    # Persist metadata
    raw_path = os.path.join(OCR_RAW_FOLDER, f"{session_id}_raw.json")
    with open(raw_path, 'w', encoding='utf-8') as f:
        json.dump({
            'session_id': session_id,
            'pdf': os.path.basename(pdf_path),
            'total_pages': total_pages,
            'chunks': total_chunks,
            'chunk_details': chunk_metadata,
            'total_chars': len(full_text),
        }, f, indent=2)

    # Persist full text
    processed_path = os.path.join(OCR_PROCESSED_FOLDER, f"{session_id}_text.txt")
    with open(processed_path, 'w', encoding='utf-8') as f:
        f.write(full_text)

    logger.info(f"OCR complete: {total_pages} pages → {len(full_text):,} chars")
    return full_text


def verify_ocr_connection() -> dict:
    """
    Verify Google Document AI credentials and API connectivity.
    Calls ListProcessors — a real API call that validates auth without uploading a document.
    Returns: { 'ok': bool, 'message': str, 'details': dict }
    """
    try:
        creds_path = _resolve_credentials()

        if not creds_path or not os.path.exists(creds_path):
            return {
                'ok': False,
                'message': 'Credentials file not found. Check GOOGLE_APPLICATION_CREDENTIALS.',
                'details': {},
            }

        # Validate credentials JSON structure
        with open(creds_path, 'r') as f:
            creds_data = json.load(f)

        required = ['type', 'project_id', 'private_key', 'client_email']
        missing = [k for k in required if k not in creds_data]
        if missing:
            return {
                'ok': False,
                'message': f'Credentials JSON missing required fields: {missing}',
                'details': {},
            }

        if creds_data.get('type') != 'service_account':
            return {
                'ok': False,
                'message': f'Expected service_account credentials, got: {creds_data.get("type")}',
                'details': {},
            }

        # Try get_processor (requires documentai.processors.get).
        # If that's also denied, fall back to confirming credentials are structurally valid —
        # a 403 means auth itself succeeded (credentials work), just limited IAM scope.
        from google.cloud import documentai_v1 as documentai
        from google.api_core.exceptions import PermissionDenied
        client = documentai.DocumentProcessorServiceClient()

        processor_path = client.processor_path(
            GOOGLE_PROJECT_ID, GOOGLE_LOCATION, GOOGLE_PROCESSOR_ID
        )
        try:
            proc = client.get_processor(
                request=documentai.GetProcessorRequest(name=processor_path)
            )
            processor_state = proc.state.name  # e.g. "ENABLED"
            status_msg = f'Processor {GOOGLE_PROCESSOR_ID} found — state: {processor_state}.'
            target_found = True
        except PermissionDenied:
            # Service account has process-only rights — this is normal.
            # Credentials are valid; we just can't call admin APIs.
            status_msg = (
                f'Credentials valid. Processor admin check skipped '
                f'(service account has process-only IAM scope — this is correct).'
            )
            target_found = True  # trust config; real failure shows at process time
        except Exception as probe_err:
            status_msg = f'Processor probe returned: {probe_err}'
            target_found = False

        return {
            'ok': True,
            'message': f'Google Document AI: credentials verified. {status_msg}',
            'details': {
                'project_id': GOOGLE_PROJECT_ID,
                'processor_id': GOOGLE_PROCESSOR_ID,
                'location': GOOGLE_LOCATION,
                'service_account': creds_data.get('client_email', ''),
                'target_processor_configured': target_found,
                'chunk_size_pages': CHUNK_SIZE,
            },
        }

    except Exception as e:
        return {
            'ok': False,
            'message': f'OCR verification failed — {type(e).__name__}: {str(e)}',
            'details': {},
        }
