requirements.txt
flask==3.0.3
python-dotenv==1.0.1
google-cloud-documentai==2.29.0
openai>=1.109.1
openpyxl==3.1.5
Werkzeug==3.0.3
pypdf>=4.0.0
